requirements.txt
alembic
bcrypt
dateparser
Flask
flask_cors
flask-login
flask_sqlalchemy
flask_script
gunicorn
hypothesis
inflect
isbnlib
nltk
pandas
PostgreSQL-Audit
psycopg2-binary
pytest
requests
simplejson
wetransferpy
xlrd
